requirements.txt
streamlit
google-generativeai
openai
pypdf2
pytesseract 
pdf2image 
pillow
python-dotenv
